+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add "language" config option, to easily change the language of in-game text.
    Thanks to the generous help of translators, the following built-in languages are available:
    CN, ES, ES2, ES3, KO, ZH
+ - Fix an issue where the vanilla /give command could duplicate items when a wand inventory is open
 
 # 7.8.3